backend/requirements.txt
# Django Backend Dependencies
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
django-cors-headers==4.6.0
django-redis==5.4.0
django-filter==24.3
django-ratelimit==4.1.0
django-otp==1.3.0
psycopg[binary]==3.2.3
argon2-cffi==23.1.0
gunicorn==23.0.0
python-decouple==3.8
dj-database-url==2.2.0
Pillow==10.4.0
celery==5.3.6
openpyxl==3.1.5
xlrd==2.0.1

# FastAPI Timetable Service Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
ortools==9.14.6206
python-constraint==1.4.0
numpy>=1.24.0
networkx==3.2.1
python-louvain==0.16
scikit-learn>=1.3.0
websockets==12.0

# Shared Dependencies
redis==5.2.0
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
sentry-sdk==2.19.2
python-jose[cryptography]==3.3.0
setuptools>=65.0.0
psutil==6.1.0  # System resource monitoring for adaptive parallelism

# Phase 1: Real-Time Progress & WebSocket
websockets==12.0  # Already present, keeping for clarity
redis[hiredis]==5.2.0  # High-performance Redis pub/sub

# Phase 2: Monitoring & Observability
prometheus-client==0.21.0  # Prometheus metrics
prometheus-fastapi-instrumentator==7.0.0  # FastAPI metrics
flower==2.0.1  # Celery monitoring (already for Celery)

# Phase 3: Object Storage & Security
boto3==1.35.76  # AWS S3 client
minio==7.2.11  # MinIO client (S3-compatible)
django-storages==1.14.4  # Django S3 integration

# Testing & Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
faker==20.1.0
locust==2.20.0
bandit==1.7.5
black==23.12.1
flake8==6.1.0
isort==5.13.2
